# tests/server/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KeyCreateRequest": {
        "description": "Request body for POST /keys (D7 BE-4).",
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          }
        },
        "required": [
          "namespace"
        ],
        "title": "KeyCreateRequest",
        "type": "object"
      },
      "KeyCreateResponse": {
        "description": "Response body for POST /keys (D7 BE-4).\n\nThe ``token`` field is present only in the create response \u2014 it is printed\nonce and never recoverable from the server thereafter.  ``status`` is always\n``'active'`` at creation time (for consistency with ``KeyResponse`` which\ncarries the live status value).\n\nSee contract C3 in ``Documentation/Backlog/D7-multi-key-auth-rotation-team-plan.md``.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "revoked"
            ],
            "title": "Status",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "token",
          "namespace",
          "created_at"
        ],
        "title": "KeyCreateResponse",
        "type": "object"
      },
      "KeyListResponse": {
        "description": "Response body for GET /keys (D7 BE-6).\n\n``hidden_revoked_count`` is the count of revoked keys excluded when the\n``status`` filter is ``'active'`` (default).  Always 0 when the filter is\n``'revoked'`` or ``'all'``.",
        "properties": {
          "hidden_revoked_count": {
            "default": 0,
            "title": "Hidden Revoked Count",
            "type": "integer"
          },
          "keys": {
            "items": {
              "$ref": "#/components/schemas/KeyResponse"
            },
            "title": "Keys",
            "type": "array"
          }
        },
        "required": [
          "keys"
        ],
        "title": "KeyListResponse",
        "type": "object"
      },
      "KeyResponse": {
        "description": "A single key record in list/revoke responses \u2014 no token field (D7 BE-6).\n\n``id`` is ``str | None`` to accommodate TOML synthetic keys (which have\n``id=None`` since they are not managed entries).  TOML synthetic keys\ncannot be targeted by ``DELETE /keys/{id}``.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "revoked"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "namespace",
          "created_at"
        ],
        "title": "KeyResponse",
        "type": "object"
      },
      "KeyRevokeResponse": {
        "description": "Response body for DELETE /keys/{id} (D7 BE-6).\n\n``status`` is always ``'revoked'`` \u2014 a successful DELETE always produces a revoked key.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "const": "revoked",
            "default": "revoked",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "KeyRevokeResponse",
        "type": "object"
      },
      "KeyRotateRequest": {
        "description": "Request body for POST /keys/rotate (D7 BE-8).\n\n``grace_seconds`` overrides the TOML ``[auth].rotate_grace_seconds`` default.\nIf absent, the TOML default is used.  If both are 0, rotation is immediate.",
        "properties": {
          "grace_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grace Seconds"
          }
        },
        "title": "KeyRotateRequest",
        "type": "object"
      },
      "KeyRotateResponse": {
        "description": "Response body for POST /keys/rotate (D7 BE-8).\n\nThe ``token`` field carries the new raw bearer token \u2014 printed once and\nnever recoverable.  Optional ``old_key_*`` fields are populated when a\nprevious default managed key was found and mutated (revoked or\ngrace-expiring).",
        "properties": {
          "new_key_id": {
            "title": "New Key Id",
            "type": "string"
          },
          "old_key_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Key Expires At"
          },
          "old_key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Key Id"
          },
          "old_key_status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "revoked"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Status of the old key after rotation.  ``'active'`` with a non-null ``old_key_expires_at`` means the key is grace-expiring \u2014 it will be rejected after ``old_key_expires_at`` but is still valid until then.  ``'revoked'`` means the old key is immediately invalid.  ``null`` means no previous managed key was found (e.g., first rotation from an auto-generated key that was never in keys.json).",
            "title": "Old Key Status"
          },
          "status": {
            "const": "active",
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "new_key_id",
          "token"
        ],
        "title": "KeyRotateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for archon-search document search and collection management",
    "title": "archon-search"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/keys": {
      "get": {
        "description": "List API keys.\n\nBy default (``status=active``) only active keys are returned; revoked keys\nare counted in ``hidden_revoked_count`` as a hint to the operator.\n\nTOML synthetic keys (``id=null``) are always included in the active view\nbecause they are always active \u2014 they can only be removed by editing\n``archon-search.toml`` and restarting the server.\n\nQuery parameters:\n\n- ``status`` \u2014 ``active`` (default), ``revoked``, or ``all``.\n- ``namespace`` \u2014 optional namespace scope applied before status filtering;\n  ``hidden_revoked_count`` reflects only revoked keys within this namespace.",
        "operationId": "list_keys_keys_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "revoked",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namespace"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Keys",
        "tags": [
          "keys"
        ]
      },
      "post": {
        "description": "Issue a new managed API key.\n\nThe raw bearer token is returned exactly once in the response ``token``\nfield. It is never stored on disk \u2014 only its SHA-256 hash is persisted.",
        "operationId": "create_key_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Key",
        "tags": [
          "keys"
        ]
      }
    },
    "/keys/rotate": {
      "post": {
        "description": "Rotate the default API key.\n\nGenerates a new managed API key, writes the new raw token to ``.search.env``,\nand revokes (or grace-expires) the old default key in ``keys.json``.\n\nReturns 409 when ``ARCHON_SEARCH_API_KEY`` env var is set \u2014 the env var\nalways overrides ``.search.env`` in the running process, so rotation would\nbe a silent no-op from the operator's perspective (S23).\n\nThe ``grace_seconds`` request body field overrides the TOML\n``[auth].rotate_grace_seconds`` config default.  When absent, the config\ndefault is used.  When both are 0, the old key is immediately revoked.",
        "operationId": "rotate_key_keys_rotate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyRotateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyRotateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Rotate Key",
        "tags": [
          "keys"
        ]
      }
    },
    "/keys/{key_id}": {
      "delete": {
        "description": "Revoke a managed API key.\n\nIdempotent: calling this on an already-revoked key returns 200 (desired\nstate already achieved).\n\nReturns 404 for unknown IDs.  TOML synthetic keys (``id=null``) cannot be\ntargeted by this endpoint \u2014 the literal string ``\"null\"`` returns 404 with\na hint to edit ``archon-search.toml`` instead.",
        "operationId": "revoke_key_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyRevokeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Revoke Key",
        "tags": [
          "keys"
        ]
      }
    },
...
  }
}